Chunk by Japanese vs non-Japanese; pass symbols/digits through
Live conversion split the composing buffer on punctuation, but a
non-Japanese run (digits/symbols/alphabet) was still fed to the neural
converter, which sometimes dropped interior digits (e.g. 123456).

Replace the punctuation-break rule with a single classification:
is_japanese (hiragana, katakana incl. ー, kanji) vs everything else. A
new chunk starts when the current one is full OR the group changes;
punctuation, being non-Japanese, now separates clauses for free. A
non-Japanese chunk is passed through to the preedit verbatim and never
sent to the model, so digits can no longer be mangled. The middle dot ・
is special-cased as a non-Japanese separator (ジョン・スミス splits),
while スーパーマーケット stays one Japanese chunk.

Extract the whole chunking subsystem (is_japanese/group_chunks/
ChunkPlan/chunked_auto_suggest + tests) into a new chunk.rs module,
shrinking conversion.rs from 1344 to 836 lines.
