+++ karukan-im/src/core/engine/mod.rs
@@ -3,6 +3,7 @@
 //! This module contains the main `InputMethodEngine` struct that coordinates between
 //! the romaji converter, kanji converter, and manages the IME state.
 
+mod chunk;
 mod conversion;
 mod cursor;
 mod display;
